Move frame resources from Scheduler to GpuExecutor
