+++ src/Teide/GpuExecutor.cpp
@@ -11,8 +11,34 @@
 namespace Teide
 {
 
-GpuExecutor::GpuExecutor(vk::Device device, vk::Queue queue) : m_device{device}, m_queue{queue}
+namespace
 {
+    void SetCommandBufferDebugName(vk::UniqueCommandBuffer& commandBuffer, uint32 threadIndex, uint32 cbIndex)
+    {
+        SetDebugName(commandBuffer, "RenderThread{}:CommandBuffer{}", threadIndex, cbIndex);
+    }
+
+} // namespace
+
+GpuExecutor::GpuExecutor(uint32 numThreads, vk::Device device, vk::Queue queue, uint32 queueFamilyIndex) :
+    m_device{device}, m_queue{queue}
+{
+    std::ranges::generate(m_frameResources, [&]() {
+        std::vector<ThreadResources> ret;
+        ret.resize(numThreads);
+        uint32 i = 0;
+        std::ranges::generate(ret, [&] {
+            ThreadResources res;
+            res.commandPool = CreateCommandPool(queueFamilyIndex, device);
+            SetDebugName(res.commandPool, "RenderThread{}:CommandPool", i);
+            res.threadIndex = i;
+            i++;
+            return res;
+        });
+
+        return ret;
+    });
+
     m_schedulerThread = std::thread([this] {
         constexpr auto timeout = std::chrono::milliseconds{2};
 
@@ -81,6 +107,58 @@ GpuExecutor::~GpuExecutor() noexcept
             spdlog::error("Timeout while waiting for command buffer execution to complete!");
         }
     }
+}
+
+void GpuExecutor::NextFrame()
+{
+    m_frameNumber = (m_frameNumber + 1) % MaxFramesInFlight;
+
+    auto& frameResources = m_frameResources[m_frameNumber];
+    for (auto& threadResources : frameResources)
+    {
+        threadResources.Reset(m_device);
+    }
+}
+
+void GpuExecutor::ThreadResources::Reset(vk::Device device)
+{
+    device.resetCommandPool(commandPool.get());
+    numUsedCommandBuffers = 0;
+
+    // Resetting also resets the command buffers' debug names
+    for (uint32 i = 0; i < commandBuffers.size(); i++)
+    {
+        commandBuffers[i].Reset();
+        SetCommandBufferDebugName(commandBuffers[i].Get(), threadIndex, i);
+    }
+}
+
+CommandBuffer& GpuExecutor::GetCommandBuffer(uint32 threadIndex)
+{
+    auto& threadResources = m_frameResources[m_frameNumber][threadIndex];
+
+    if (threadResources.numUsedCommandBuffers == threadResources.commandBuffers.size())
+    {
+        const vk::CommandBufferAllocateInfo allocateInfo = {
+            .commandPool = threadResources.commandPool.get(),
+            .level = vk::CommandBufferLevel::ePrimary,
+            .commandBufferCount = std::max(1u, static_cast<uint32>(threadResources.commandBuffers.size())),
+        };
+
+        auto newCommandBuffers = m_device.allocateCommandBuffersUnique(allocateInfo);
+        const auto numCBs = static_cast<uint32>(threadResources.commandBuffers.size());
+        for (uint32 i = 0; i < newCommandBuffers.size(); i++)
+        {
+            SetCommandBufferDebugName(newCommandBuffers[i], threadIndex, i + numCBs);
+            threadResources.commandBuffers.emplace_back(std::move(newCommandBuffers[i]));
+        }
+    }
+
+    const auto commandBufferIndex = threadResources.numUsedCommandBuffers++;
+    CommandBuffer& commandBuffer = threadResources.commandBuffers.at(commandBufferIndex);
+    commandBuffer.Get()->begin(vk::CommandBufferBeginInfo{});
+
+    return commandBuffer;
 }
 
 uint32 GpuExecutor::AddCommandBufferSlot()
