+++ tests/TeideTest/src/Teide/GpuExecutorTest.cpp
@@ -46,6 +46,11 @@ protected:
     vk::Device GetDevice() const { return m_device.get(); }
     vk::Queue GetQueue() const { return m_queue; }
 
+    GpuExecutor CreateGpuExecutor()
+    {
+        return {2, GetDevice(), GetQueue(), m_physicalDevice.queueFamilies.transferFamily};
+    }
+
     vk::UniqueCommandBuffer CreateCommandBuffer()
     {
         auto list = m_device->allocateCommandBuffersUnique({.commandPool = m_commandPool.get(), .commandBufferCount = 1});
@@ -82,7 +87,7 @@ private:
 
 TEST_F(GpuExecutorTest, OneCommandBuffer)
 {
-    auto executor = GpuExecutor(GetDevice(), GetQueue());
+    auto executor = CreateGpuExecutor();
     auto cmdBuffer = CreateCommandBuffer();
     auto buffer = CreateHostVisibleBuffer(12);
 
@@ -102,7 +107,7 @@ TEST_F(GpuExecutorTest, OneCommandBuffer)
 
 TEST_F(GpuExecutorTest, TwoCommandBuffers)
 {
-    auto executor = GpuExecutor(GetDevice(), GetQueue());
+    auto executor = CreateGpuExecutor();
     auto cmdBuffer1 = CreateCommandBuffer();
     auto cmdBuffer2 = CreateCommandBuffer();
     auto buffer = CreateHostVisibleBuffer(12);
@@ -130,7 +135,7 @@ TEST_F(GpuExecutorTest, TwoCommandBuffers)
 
 TEST_F(GpuExecutorTest, TwoCommandBuffersOutOfOrder)
 {
-    auto executor = GpuExecutor(GetDevice(), GetQueue());
+    auto executor = CreateGpuExecutor();
     auto cmdBuffer1 = CreateCommandBuffer();
     auto cmdBuffer2 = CreateCommandBuffer();
     auto buffer = CreateHostVisibleBuffer(12);
@@ -161,7 +166,7 @@ TEST_F(GpuExecutorTest, DontWait)
     auto buffer = CreateHostVisibleBuffer(12);
     auto cmdBuffer = CreateCommandBuffer();
     {
-        auto executor = GpuExecutor(GetDevice(), GetQueue());
+        auto executor = CreateGpuExecutor();
 
         cmdBuffer->begin(vk::CommandBufferBeginInfo{});
         cmdBuffer->fillBuffer(buffer.buffer.get(), 0, 12, 0x01010101);
@@ -174,7 +179,7 @@ TEST_F(GpuExecutorTest, DontWait)
 
 TEST_F(GpuExecutorTest, SubmitMultipleCommandBuffers)
 {
-    auto executor = GpuExecutor(GetDevice(), GetQueue());
+    auto executor = CreateGpuExecutor();
     for (int i = 0; i < 2; i++)
     {
         auto cmdBuffer = CreateCommandBuffer();
